Extract `RequiredFieldGenerator`
